Create modal.components page
